requirements.txt
PyYAML == 5.3.1
addict == 2.4.0
albumentations == 1.0.3
numpy == 1.19.1
opencv-python == 4.4.0.46
pandas == 1.1.1
torch == 1.6.0
tqdm == 4.47.0
Pillow == 7.2.0
seaborn